Add support for JsonPropertyNames and hardcode Enumerations fix
Add support for JsonPropertyNames that are added to DTOs to override their given property name from this project.

We use a Enumerations base class that contains all enums, but it was not mapping correctly so I have just hardcoded a fix for now where we remove the "Enumerations." value from a property type when referring to Enumerations.